# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "RakursPRO",
        "description": "API документация для проекта RakursPRO",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Local server"
        }
    ],
    "paths": {
...
        "/pages": {
            "get": {
                "tags": [
                    "Страницы"
                ],
                "summary": "Получить список всех страниц",
                "description": "Возвращает массив всех доступных страниц с их идентификаторами, названиями и слагами. Требуется авторизация.",
                "operationId": "27d81f8e8c107b6dadce528d16473a15",
                "parameters": [
                    {
                        "name": "Accept",
                        "in": "header",
                        "description": "Тип принимаемого контента",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "application/json"
                        }
                    },
                    {
                        "name": "Content-Type",
                        "in": "header",
                        "description": "Тип отправляемого контента",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "application/json"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ. Возвращает массив страниц.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Page"
                                    }
                                },
                                "example": [
                                    {
                                        "id": 1,
                                        "name": "Главная",
                                        "slug": null
                                    },
                                    {
                                        "id": 2,
                                        "name": "Макияж",
                                        "slug": "makeup"
                                    }
                                ]
                            }
                        }
                    },
                    "401": {
                        "description": "Неавторизован - отсутствует или недействителен токен доступа",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Forbidden"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Server error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Page": {
                "properties": {
                    "id": {
                        "description": "Уникальный идентификатор страницы",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Название страницы",
                        "type": "string",
                        "example": "Главная"
                    },
                    "slug": {
                        "description": "URL-слаг страницы (может быть null)",
                        "type": "string",
                        "example": "makeup",
                        "nullable": true
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Введите токен в формате: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}